requirements.txt
fastapi
uvicorn
google-generativeai
python-telegram-bot
requests
pillow
edge-tts

